pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-trial-mas"
version = "0.2.1"
description = "Multi-Agent AI for Clinical Trial Data Management — LangGraph reference implementation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "DanielMartin Arogyasami"}]
keywords = ["clinical-trials", "multi-agent", "langgraph", "CDISC", "SDTM"]

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pyarrow>=15.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pypdf>=4.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*", "evaluation*", "data*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
